requirements.txt
certifi==2019.9.11
chardet==3.0.4
idna==2.8
mock==3.0.5
Pillow==6.1.0
praw==6.3.1
prawcore==1.0.1
requests==2.22.0
six==1.12.0
update-checker==0.16
urllib3==1.25.3
websocket-client==0.56.0
